Initial Strassen
Still have not worked on this. This is just copied from https://martin-thoma.com/strassen-algorithm-in-python-java-cpp/.
We need to talk about how we want to structure our project. We need to discuss how we are planning on creating the matrix (random or user generated), then where the zero-filler should go. I suggest in the same file. The basic multiplication should be its own file and the Strassen algorithm should be its own file. I apologize for not working on this Saturday. Busy weekend with homework and Deck the Halls.